ngerapihin file onboarding view, dan fix bug gak bisa start exercise abis allow microphone di onboarding permission(baru bisa exercise ketika restart app), maka ketika user allow microphone permission aplikasi otomatis ter-restart
